Update README.md
This program is designed to open and read an input file which contains an unknown number of records. It is then to print the list of first names, last names, and test scores. It will then sort the list by last name from A to Z and print. Then the program will calculate the average test score for each set and determine a letter grade based on it and print. The program will then sort the lists based on test scores from high to low and print the final list.